del: unnecessary optimizers
